+++ packages/react-pdf-toolkit/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ballerine/react-pdf-toolkit
 
+## 1.2.5
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @ballerine/config@1.1.4
+  - @ballerine/ui@0.5.5
+
+## 1.2.4
+
+### Patch Changes
+
+- document changes
+- Updated dependencies
+  - @ballerine/config@1.1.3
+  - @ballerine/ui@0.5.4
+
 ## 1.2.3
 
 ### Patch Changes