+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @ballerine/ui
 
+## 0.5.5
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+- Updated dependencies
+  - @ballerine/common@0.9.11
+
+## 0.5.4
+
+### Patch Changes
+
+- document changes
+- Updated dependencies
+  - @ballerine/common@0.9.10
+
 ## 0.5.3
 
 ### Patch Changes