+++ services/workflows-service/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @ballerine/workflows-service
 
+## 0.7.16
+
+### Patch Changes
+
+- Bump
+- Bump
+- Updated dependencies
+- Updated dependencies
+  - @ballerine/common@0.9.11
+  - @ballerine/workflow-core@0.6.17
+  - @ballerine/workflow-node-sdk@0.6.17
+
+## 0.7.15
+
+### Patch Changes
+
+- document changes
+- Updated dependencies
+  - @ballerine/common@0.9.10
+  - @ballerine/workflow-core@0.6.16
+  - @ballerine/workflow-node-sdk@0.6.16
+
+## 0.7.14
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.9
+  - @ballerine/workflow-core@0.6.15
+  - @ballerine/workflow-node-sdk@0.6.15
+
 ## 0.7.13
 
 ### Patch Changes